requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via pip-tools
django==5.1.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
joblib==1.4.2
    # via scikit-learn
numpy==2.2.2
    # via
    #   scikit-learn
    #   scipy
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.1
    # via scikit-learn
sqlparse==0.5.3
    # via django
threadpoolctl==3.5.0
    # via scikit-learn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
